pyproject.toml
[tool.poetry]
name = "chat-parsing"
version = "0.1.0"
description = ""
authors = ["zewa-bamuga <tikhonov.igor2028@yandex.ru>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.36.0"
asyncio = "^3.4.3"
python-dotenv = ">=0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
